// frontend/src/app/gis/shared/PropertyLayer.enum.ts
export enum PropertyLayer {

    /* Nome delle proprieta dei Layer definite su tabella A4ST_PROPRIETA_LAYER */
    TITOLO = 'titolo',
    CANCELLABILE = 'cancellabile',
    TEMATISMOLABEL = 'tematismoLabel',
    POSIZIONE = 'posizione',
    ABILITATODEFAULT = 'AbilitatoDefault',
    TEMATISMOBORDOCOLORE = 'tematismoBordoColore',
    TEMATISMOFILLCOLORE = 'tematismoFillColore',
    TEMATISMOEVIDENTE = 'tematismoEvidente',
    TEMATISMOBORDOSPESSORE = 'TEMATISMOBORDOSPESSORE',
    FORMATO = 'formato',
    CONTESTO = 'contesto',
    ISPEZIONABILE = 'Ispezionabile',
    GRUPPO = 'gruppo',
    COLLASSAGRUPPO = 'collassaGruppo',
    CAMPOFILTRO = 'campoFiltro',
    EDITABILE = 'editabile',
    TIPO_WMS = 'TIPO_WMS',
    CONTENUTO_INFORMATIVO = 'CONTENUTO_INFORMATIVO',
    FORMATO_IMMAGINE = 'FORMATO_IMMAGINE',
    ATTRIBUTION = 'ATTRIBUTION',
    CAMPO_IDENTIFY = 'CAMPO_IDENTIFY',
    SNAP = 'snap',
    MAX_SCALE = 'MAX_SCALE',
    MIN_SCALE = 'MIN_SCALE',
    TILE_SIZE = 'TILE_SIZE',
    DISPLAYINLAYERSWITCHER = 'displayInLayerSwitcher',
    RELAZIONECONLAYER = 'relazioneConLayer',
    SPECIFICO_PER_CAMPAGNA = 'SPECIFICO_PER_CAMPAGNA',
    CQL_CAMPAGNA = 'CQL_CAMPAGNA',
    WFS_STYLE = 'WFS_STYLE',
    WFS_STYLE_CLIP_SU_ADL = 'CLIP_SU_ADL',
    WFS_STYLE_ANOMALIE = 'ANOMALIE',
    CARICAMENTO_STORICO = 'CARICAMENTO_STORICO',

    /* Valore  delle proprieta dei Layer definite su tabella A4ST_PROPRIETA_LAYER */
    FORMATO_WFS = 'wfs',
    FORMATO_WMS = 'wms',
    TIPO_WMS_TILE = 'Tile',
    TIPO_WMS_IMAGE = 'Image',
    TIPO_WMS_WMTS = 'Wmts',
    SI = 'si',
    NO = 'no',
    CONTESTO_LAVORAZIONE_EDITING = 'LavorazioneEditing',
    CONTENUTO_INFORMATIVO_ISTANZA_DI_RIESAME = 'ISTANZA_DI_RIESAME',
    CONTENUTO_INFORMATIVO_WORKSPACE = 'WORKSPACE_LAVORAZIONE',
    CONTENUTO_INFORMATIVO_SUOLO_OCCUPATO = 'SUOLO_IN_LAVORAZIONE',
    CODICE_LAYER_WORKSPACE = 'workspaceBO',
    CODICE_LAYER_ISTANZA_DI_RIESAME_BO = 'IstanzaDiRiesameBO',
    CODICE_LAYER_ISTANZA_DI_RIESAME_LAV_BO = 'IstanzaDiRiesameLavBO',
    CODICE_LAYER_SUOLO_PRENOTATO_LAV_BO = 'SuoloPrenotatoLavBO',
    CODICE_LAYER_SUOLO_VIGENTE_BO = 'suoloVigenteBO',
    CODICE_LAYER_SUOLO_STORICO = 'suoloStorico',
    NOME_LAYER_SUOLO_STORICO = 'A4SV_SUOLO_STORICO_LAYER2',
    CODICE_LAYER_SUOLO_VIGENTE_INCLUSO_IN_LAVORAZIONI_BO = 'suoloVigenteInclusoInLavorazioniBO',
    CODICE_LAYER_ANOMALIE_VALIDAZIONE_BO = 'anomalieValidazioneBO',
    CODICE_LAYER_CLIP_SU_ADL = 'TempClipSuAdl',
    CODICE_LAYER_ADL = 'AreaDilavoro',
    NOME_LAYER_TMP = 'NOME_LAYER_TMP',
    FORMATO_IMAGE_DEFAULT = 'image/png8',

    /* Altro */
    VALOREFILTRO = 'valoreFiltro',
    TITLE = 'title',
    WORKSPACE = 'WORKSPACE',
    LABEL = 'label',
    CODICE = 'codice',
    NOMELAYER = 'nomeLayer',
    NOSWITCHERDELETE = 'noSwitcherDelete',
    ALLWAYSONTOP = 'allwaysOnTop',
    VISIBLE = 'visible',
    OPENINLAYERSWITCHER = 'openInLayerSwitcher',
    PARAMS_FILTER = 'PARAMS_FILTER',
    STYLE = 'style',
}